Implement /collections/{id}/thumbnail route
