+++ test/fixtures/mocks/missing-collection-1234.json
@@ -1,0 +1,6 @@
+{
+  "_index": "dev-dc-v2-collection",
+  "_type": "_doc",
+  "_id": "1234",
+  "found": false
+}
