+++ test/integration/get-thumbnail.test.js
@@ -2,124 +2,224 @@
 
 const chai = require("chai");
 const expect = chai.expect;
-const { handler } = require("../../src/handlers/get-work-thumbnail");
+const { handler } = require("../../src/handlers/get-thumbnail");
 
-describe("Work thumbnail", () => {
+describe("Thumbnail routes", () => {
   helpers.saveEnvironment();
   const mock = helpers.mockIndex();
-  const event = helpers
-    .mockEvent("GET", "/works/{id}/thumbnail")
-    .pathPrefix("/api/v2")
-    .pathParams({ id: 1234 });
 
   beforeEach(() => {
     process.env.API_TOKEN_SECRET = "abcdef";
   });
 
-  it("retrieves a thumbnail", async () => {
-    mock
-      .get("/dc-v2-work/_doc/1234")
-      .reply(200, helpers.testFixture("mocks/work-1234.json"));
-    mock
-      .get("/iiif/2/mbk-dev/5678/full/!300,300/0/default.jpg")
-      .reply(200, helpers.testFixture("mocks/thumbnail_full.jpg"), {
-        "Content-Type": "image/jpeg",
-      });
+  describe("Collection", () => {
+    const event = helpers
+      .mockEvent("GET", "/collections/{id}/thumbnail")
+      .pathPrefix("/api/v2")
+      .pathParams({ id: 1234 });
 
-    const result = await handler(event.render());
-    expect(result.statusCode).to.eq(200);
-    expect(result.headers["content-type"]).to.eq("image/jpeg");
+    it("retrieves a thumbnail", async () => {
+      mock
+        .get("/dc-v2-collection/_doc/1234")
+        .reply(200, helpers.testFixture("mocks/collection-1234.json"));
+      mock
+        .get("/dc-v2-work/_doc/1234")
+        .reply(200, helpers.testFixture("mocks/work-1234.json"));
+      mock
+        .get("/iiif/2/mbk-dev/5678/full/!300,300/0/default.jpg")
+        .reply(200, helpers.testFixture("mocks/thumbnail_full.jpg"), {
+          "Content-Type": "image/jpeg",
+        });
+
+      const result = await handler(event.render());
+      expect(result.statusCode).to.eq(200);
+      expect(result.headers["content-type"]).to.eq("image/jpeg");
+    });
+
+    it("returns an error from the IIIF server", async () => {
+      mock
+        .get("/dc-v2-collection/_doc/1234")
+        .reply(200, helpers.testFixture("mocks/collection-1234.json"));
+      mock
+        .get("/dc-v2-work/_doc/1234")
+        .reply(200, helpers.testFixture("mocks/work-1234.json"));
+      mock
+        .get("/iiif/2/mbk-dev/5678/full/!300,300/0/default.jpg")
+        .reply(403, "Forbidden", { "Content-Type": "text/plain" });
+
+      const result = await handler(event.render());
+      expect(result.statusCode).to.eq(403);
+      expect(result.body).to.eq("Forbidden");
+    });
+
+    it("returns 404 if the collection doc can't be found", async () => {
+      mock
+        .get("/dc-v2-collection/_doc/1234")
+        .reply(200, helpers.testFixture("mocks/missing-collection-1234.json"));
+
+      const result = await handler(event.render());
+      expect(result.statusCode).to.eq(404);
+    });
+
+    it("returns 404 if the work doc can't be found", async () => {
+      mock
+        .get("/dc-v2-collection/_doc/1234")
+        .reply(200, helpers.testFixture("mocks/collection-1234.json"));
+      mock
+        .get("/dc-v2-work/_doc/1234")
+        .reply(200, helpers.testFixture("mocks/missing-work-1234.json"));
+
+      const result = await handler(event.render());
+      expect(result.statusCode).to.eq(404);
+    });
+
+    it("returns 404 if the collection doc has no representative work", async () => {
+      mock
+        .get("/dc-v2-collection/_doc/1234")
+        .reply(
+          200,
+          helpers.testFixture("mocks/collection-1234-no-thumbnail.json")
+        );
+
+      const result = await handler(event.render());
+      expect(result.statusCode).to.eq(404);
+    });
+
+    it("returns 404 if the work doc has no thumbnail", async () => {
+      mock
+        .get("/dc-v2-collection/_doc/1234")
+        .reply(200, helpers.testFixture("mocks/collection-1234.json"));
+      mock
+        .get("/dc-v2-work/_doc/1234")
+        .reply(200, helpers.testFixture("mocks/work-1234-no-thumbnail.json"));
+
+      const result = await handler(event.render());
+      expect(result.statusCode).to.eq(404);
+    });
   });
 
-  it("returns an error from the IIIF server", async () => {
-    mock
-      .get("/dc-v2-work/_doc/1234")
-      .reply(200, helpers.testFixture("mocks/work-1234.json"));
-    mock
-      .get("/iiif/2/mbk-dev/5678/full/!300,300/0/default.jpg")
-      .reply(403, "Forbidden", { "Content-Type": "text/plain" });
+  describe("Work", () => {
+    const event = helpers
+      .mockEvent("GET", "/works/{id}/thumbnail")
+      .pathPrefix("/api/v2")
+      .pathParams({ id: 1234 });
 
-    const result = await handler(event.render());
-    expect(result.statusCode).to.eq(403);
-    expect(result.body).to.eq("Forbidden");
+    it("retrieves a thumbnail", async () => {
+      mock
+        .get("/dc-v2-work/_doc/1234")
+        .reply(200, helpers.testFixture("mocks/work-1234.json"));
+      mock
+        .get("/iiif/2/mbk-dev/5678/full/!300,300/0/default.jpg")
+        .reply(200, helpers.testFixture("mocks/thumbnail_full.jpg"), {
+          "Content-Type": "image/jpeg",
+        });
+
+      const result = await handler(event.render());
+      expect(result.statusCode).to.eq(200);
+      expect(result.headers["content-type"]).to.eq("image/jpeg");
+    });
+
+    it("returns an error from the IIIF server", async () => {
+      mock
+        .get("/dc-v2-work/_doc/1234")
+        .reply(200, helpers.testFixture("mocks/work-1234.json"));
+      mock
+        .get("/iiif/2/mbk-dev/5678/full/!300,300/0/default.jpg")
+        .reply(403, "Forbidden", { "Content-Type": "text/plain" });
+
+      const result = await handler(event.render());
+      expect(result.statusCode).to.eq(403);
+      expect(result.body).to.eq("Forbidden");
+    });
+
+    it("returns 404 if the work doc can't be found", async () => {
+      mock
+        .get("/dc-v2-work/_doc/1234")
+        .reply(200, helpers.testFixture("mocks/missing-work-1234.json"));
+
+      const result = await handler(event.render());
+      expect(result.statusCode).to.eq(404);
+    });
+
+    it("returns 404 if the work doc has no thumbnail", async () => {
+      mock
+        .get("/dc-v2-work/_doc/1234")
+        .reply(200, helpers.testFixture("mocks/work-1234-no-thumbnail.json"));
+
+      const result = await handler(event.render());
+      expect(result.statusCode).to.eq(404);
+    });
   });
 
-  it("returns 404 if the work doc can't be found", async () => {
-    mock
-      .get("/dc-v2-work/_doc/1234")
-      .reply(200, helpers.testFixture("mocks/missing-work-1234.json"));
+  describe("QueryString parameters", () => {
+    const event = helpers
+      .mockEvent("GET", "/works/{id}/thumbnail")
+      .pathPrefix("/api/v2")
+      .pathParams({ id: 1234 });
 
-    const result = await handler(event.render());
-    expect(result.statusCode).to.eq(404);
-  });
+    it("accepts a proper size", async () => {
+      mock
+        .get("/dc-v2-work/_doc/1234")
+        .reply(200, helpers.testFixture("mocks/work-1234.json"));
+      mock
+        .get("/iiif/2/mbk-dev/5678/full/!200,200/0/default.jpg")
+        .reply(200, helpers.testFixture("mocks/thumbnail_full.jpg"), {
+          "Content-Type": "image/jpeg",
+        });
 
-  it("returns 404 if the work doc has no thumbnail", async () => {
-    mock
-      .get("/dc-v2-work/_doc/1234")
-      .reply(200, helpers.testFixture("mocks/work-1234-no-thumbnail.json"));
+      const result = await handler(event.queryParams({ size: 200 }).render());
+      expect(result.statusCode).to.eq(200);
+    });
 
-    const result = await handler(event.render());
-    expect(result.statusCode).to.eq(404);
-  });
+    it("rejects invalid sizes", async () => {
+      let result = await handler(event.queryParams({ size: "foo" }).render());
+      expect(result.statusCode).to.eq(400);
+      expect(result.body).to.contain("foo is not");
 
-  it("accepts a proper size", async () => {
-    mock
-      .get("/dc-v2-work/_doc/1234")
-      .reply(200, helpers.testFixture("mocks/work-1234.json"));
-    mock
-      .get("/iiif/2/mbk-dev/5678/full/!200,200/0/default.jpg")
-      .reply(200, helpers.testFixture("mocks/thumbnail_full.jpg"), {
-        "Content-Type": "image/jpeg",
-      });
+      result = await handler(event.queryParams({ size: 500 }).render());
+      expect(result.statusCode).to.eq(400);
+      expect(result.body).to.contain("500px");
+    });
 
-    const result = await handler(event.queryParams({ size: 200 }).render());
-    expect(result.statusCode).to.eq(200);
-  });
+    it("accepts proper aspect ratios", async () => {
+      mock
+        .get("/dc-v2-work/_doc/1234")
+        .times(2)
+        .reply(200, helpers.testFixture("mocks/work-1234.json"));
 
-  it("rejects invalid sizes", async () => {
-    let result = await handler(event.queryParams({ size: "foo" }).render());
-    expect(result.statusCode).to.eq(400);
-    expect(result.body).to.contain("foo is not");
+      let resultFixture = "mocks/thumbnail_full.jpg";
+      mock
+        .get("/iiif/2/mbk-dev/5678/full/!300,300/0/default.jpg")
+        .reply(200, helpers.testFixture(resultFixture), {
+          "Content-Type": "image/jpeg",
+        });
 
-    result = await handler(event.queryParams({ size: 500 }).render());
-    expect(result.statusCode).to.eq(400);
-    expect(result.body).to.contain("500px");
-  });
+      let result = await handler(
+        event.queryParams({ aspect: "full" }).render()
+      );
+      expect(result.statusCode).to.eq(200);
+      let expected = helpers.encodedFixture(resultFixture);
+      expect(result.body).to.eq(expected);
 
-  it("accepts proper aspect ratios", async () => {
-    mock
-      .get("/dc-v2-work/_doc/1234")
-      .times(2)
-      .reply(200, helpers.testFixture("mocks/work-1234.json"));
+      resultFixture = "mocks/thumbnail_square.jpg";
+      mock
+        .get("/iiif/2/mbk-dev/5678/square/!300,300/0/default.jpg")
+        .reply(200, helpers.testFixture(resultFixture), {
+          "Content-Type": "image/jpeg",
+        });
 
-    let resultFixture = "mocks/thumbnail_full.jpg";
-    mock
-      .get("/iiif/2/mbk-dev/5678/full/!300,300/0/default.jpg")
-      .reply(200, helpers.testFixture(resultFixture), {
-        "Content-Type": "image/jpeg",
-      });
+      result = await handler(event.queryParams({ aspect: "square" }).render());
+      expect(result.statusCode).to.eq(200);
+      expected = helpers.encodedFixture(resultFixture);
+      expect(result.body).to.eq(expected);
+    });
 
-    let result = await handler(event.queryParams({ aspect: "full" }).render());
-    expect(result.statusCode).to.eq(200);
-    let expected = helpers.encodedFixture(resultFixture);
-    expect(result.body).to.eq(expected);
-
-    resultFixture = "mocks/thumbnail_square.jpg";
-    mock
-      .get("/iiif/2/mbk-dev/5678/square/!300,300/0/default.jpg")
-      .reply(200, helpers.testFixture(resultFixture), {
-        "Content-Type": "image/jpeg",
-      });
-
-    result = await handler(event.queryParams({ aspect: "square" }).render());
-    expect(result.statusCode).to.eq(200);
-    expected = helpers.encodedFixture(resultFixture);
-    expect(result.body).to.eq(expected);
-  });
-
-  it("rejects improper aspect ratio", async () => {
-    const result = await handler(event.queryParams({ aspect: "foo" }).render());
-    expect(result.statusCode).to.eq(400);
-    expect(result.body).to.contain("Unknown aspect ratio: foo");
+    it("rejects improper aspect ratio", async () => {
+      const result = await handler(
+        event.queryParams({ aspect: "foo" }).render()
+      );
+      expect(result.statusCode).to.eq(400);
+      expect(result.body).to.contain("Unknown aspect ratio: foo");
+    });
   });
 });
